+++ tntfl/hooks/utils.py
@@ -1,5 +1,5 @@
 import requests
-import urlparse
+import urllib.parse
 
 
 def postMattermost(mattermostUrl, apiKey, tntflUrl, game, fields, colour):
@@ -16,10 +16,10 @@ def postMattermost(mattermostUrl, apiKey, tntflUrl, game, fields, colour):
                 'pretext': '[{}]({})'.format(title, gameUrl),
                 'fields': fields,
                 'color': colour,
-            }
+            },
         ],
         'username': 'ScoreBot',
-        'icon_url': urlparse.urljoin(mattermostUrl, '/static/emoji/26bd.png'),
+        'icon_url': urllib.parse.urljoin(mattermostUrl, '/static/emoji/26bd.png'),
     }
-    webhookUrl = urlparse.urljoin(urlparse.urljoin(mattermostUrl, '/hooks/'), apiKey)
+    webhookUrl = urllib.parse.urljoin(urllib.parse.urljoin(mattermostUrl, '/hooks/'), apiKey)
     requests.post(webhookUrl, json=message)