+++ tntfl/transforms/elo.py
@@ -6,7 +6,7 @@ class Player(object):
 def calculateSkillChange(red, blue, redScore, blueScore):
     predict = 1 / (1 + 10 ** ((red.elo - blue.elo) / 180))
     goals = blueScore + redScore
-    result = float(blueScore) / goals
+    result = blueScore / goals
     delta = (2.5 * goals) * (result - predict)
     return delta