+++ ui/components/player/box-plot.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 const ReactHighcharts = require('react-highcharts');
 require('highcharts/highcharts-more')(ReactHighcharts.Highcharts);
 const computeBoxplotStats = require('react-boxplot/dist/stats');
+const ContainerDimensions = require('react-container-dimensions').default;
 
 interface Props {
   data: number[];
@@ -37,11 +38,18 @@ export default function BoxPlot(props: Props): JSX.Element {
   };
 
   return (
-    <div>
-      {data.length >= 5
-        ? <ReactHighcharts config={config} />
-        : 'Not enough data'
-      }
-    </div>
+    <ContainerDimensions>
+      {({height}: {height: number}) => {
+        (config.chart as any).height = height - 50;
+        return (
+          <div style={{ width: 120, height: '100%', margin: -15 }}>
+            {data.length >= 5
+              ? <ReactHighcharts config={config}/>
+              : 'Not enough data'
+            }
+          </div>
+        );
+      }}
+    </ContainerDimensions>
   );
 }